+++ src/d3d11/D3D11Shader.cpp
@@ -121,7 +121,7 @@ STDMETHODIMP D3D11Shader::D3D11Include::Open(THIS_ D3D_INCLUDE_TYPE IncludeType,
 		ETERNAL_ASSERT(false);
 	}
 
-	*pBytes = IncludedFile.tellg();
+	*pBytes = (UINT)IncludedFile.tellg();
 	*ppData = new char[*pBytes];
 	IncludedFile.seekg(0, ios::beg);
 	IncludedFile.read((char*)*ppData, *pBytes);
